requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# WebSocket support
websockets==12.0

# System monitoring
psutil==5.9.6

# LLM integration (M1 Mac optimized)
# Note: llama-cpp-python may require manual installation on some systems
# Install with: pip install llama-cpp-python
llama-cpp-python>=0.2.20

# Async support
asyncio-mqtt==0.13.0

# Data handling
pydantic==2.5.0
python-multipart==0.0.6

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
python-dotenv==1.0.0